+++ packages/reactabular-table/src/index.js
@@ -2,4 +2,6 @@ export { default as Provider } from './provider';
 export { default as Header } from './header';
 export { default as Body } from './body';
 export { default as BodyRow } from './body-row';
+export { default as evaluateFormatters } from './evaluate-formatters';
+export { default as evaluateTransforms } from './evaluate-transforms';