+++ packages/reactabular-utils/CHANGELOG.md
@@ -2,12 +2,13 @@
 ==================
 
   * Feature - Add `columnsAreEqual` checker.
-  * Feature - `evaluateTransforms` will throw if all transforms aren't functions.
   * Breaking - Move `utils.countRowSpan` has been dropped as it's not needed here anymore (moved to `table-resolver`).
   * Breaking - Drop `utils.mergeClassNames`. This was replaced by [classnames](https://www.npmjs.org/package/classnames) internally and the column definition accepts formats used by `classnames`.
   * Breaking - Generalize `utils.mergePropPair` as `utils.mergeProps`. It accepts an arbitrary amount of prop collections now.
   * Breaking - Drop `utils.resolveBodyColumns`, `utils.resolveHeaderRows`, `utils.resolveRowKey`. The functionality was moved to the `table-resolver` package.
   * Breaking - Drop `utils.rowsAreEqual`. Same as `lodash.isEqual(oldRows, newRows)`.
+  * Feature - `evaluateTransforms` will throw if all transforms aren't functions.
+  * Breaking - Move `utils.evaluateFormatters` and `utils.evaluateTransforms` to `reactabular-table`. Those aren't used elsewhere so that makes a lot of sense.
 
 7.0.0 / 2016-11-03
 ==================